moved music bot-related function from tools.js to music.js
