+++ tools.js
@@ -1,29 +1,4 @@
-const shared = require('./shared.js'),
-    ytdl = require('ytdl-core'),
-    streamOptions = { seek: 0, volume: 1 };
-
-let sendMusicEmbed = (message, music, bot) => {
-    let time = {};
-    time = tools.getTimeFormat(music.duration);
-    message.channel.send('', {embed : {
-        color: 65399,
-        author: {
-            name: `BOT ${message.guild.name} : MODE DJ`,
-            icon_url: bot.user.avatarURL
-        },
-        title: 'Now playing :',
-        fields: [{
-            name: 'Titre : ',
-            value: `${music.title}`
-        }, {
-            name: 'Durée :',
-            value: `${time.hours} heures ${time.minutes} minutes et ${time.seconds} secondes`
-        }, {
-            name: 'Proposée par :',
-            value: `${music.author}`
-        }]
-    }});
-}
+const shared = require('./shared.js');
 
 module.exports = {
     getRoles : id => {
@@ -96,22 +71,5 @@ Function that check if the user that issued a message is admin or not.
                 str += " ";
         }
         return str;
-    },
-
-    playMusic : (connection, message) => {
-        console.log('1');
-        let server = shared.musicQueues[message.guild.id];
-        sendMusicEmbed(message, server.queue[0], bot);
-        if (!server.dispatcher)
-            server.dispatcher = connection.playStream(ytdl(server.queue[0].url, {filter : 'audioonly'}), streamOptions);
-        server.dispatcher.on('end', (end) => {
-            console.log('2');
-            server.dispatcher = null;
-            server.queue.shift();
-            if (server.queue[0]) {
-                console.log('3');
-                module.exports.playMusic(connection, message);
-            }
-        });
     }
 }
